Don't mutate cached data in DatasetInfoPlugin
Multiple requests to the /info endpoint are failing.
The DatasetInfoPlugin is changing the cached
metadata by removing the `_ARRAY_DIMENSIONS`
attribute key which is required on subsequent calls. I did
not dive into the caching internals to figure out why the
the value is mutable at all. If it is a simple dict on the
backend, the returned cache values could be being
changed inside the cache dict itself.

Added a test that fails without the change!