Match DatasetInfoPlugin name and entrypoint name
The DatasetInfoPlugin plugin will end up with two different names when
it is registered with its two "default" methods: by entrypoint and by
creating a DatasetInfoPlugin() object with no arguments. Having these
two fields equal helps downstream users who want to override the default
plugins. Without this, the users will have to know what the
entrypoint name is for discovered plugins so they can override the
correct plugin. By standardizing on the name of the plugin, it can be
inferred.