Support reading scalar variables through the zarr plugin (#192)
Reading scalar variables through the zarr plugin is erroring with a
ValueError: Invalid chunk_id for numpy array: 0. Should have been: ().
This is due to ((0,) * da.ndim) being () when ndims is zero. In turn, that
does not match the requested chunk (0,). This removes the invalid chunk_id
check when dealing with scalar variables (ndim = 0). The added test fails
without the change.